Добавили TaskController и перенесли туда отрисовку одной задачи
